bracket balancer++

+++ src/main/java/io/github/gh0stinthesh311/domain/Table.java
@@ -12,6 +12,7 @@ import java.util.TreeMap;
 
 import static io.github.gh0stinthesh311.utils.Formatter.wrapWithQuotes;
 import static io.github.gh0stinthesh311.utils.StringUtils.extractContentBetweenParentheses;
+import static io.github.gh0stinthesh311.utils.StringUtils.validateParentheses;
 
 public class Table {
     String name;
@@ -27,16 +28,19 @@ public class Table {
         // to do add validation for empty column definitions - "CREATE TABLE ninjas;" is invalid
         LogUtil.info("Creating columns for " + wrapWithQuotes(table.getName()));
         String extractedColumnDefinitions = extractColumnDefinitions(SQL);
-        if (extractedColumnDefinitions.isEmpty()) {
-            LogUtil.info("Column definitions are empty:" + wrapWithQuotes(extractedColumnDefinitions) + ". Define at least one column ");
+        if (validateParentheses(SQL) && !extractedColumnDefinitions.isEmpty()) {
+//            String extractedColumnDefinitions = extractColumnDefinitions(SQL);
+//            if (extractedColumnDefinitions.isEmpty()) {
+            LogUtil.info("Column definitions are empty or not provided:" + wrapWithQuotes(extractedColumnDefinitions) + ". Define at least one column.");
             Memory.getInstance().getCurrentDatabase().dropTable(table.getName());
-        } else {
+            } else {
             String[] columnDefinitions = extractedColumnDefinitions.split(",");
             for (String columnDefinition : columnDefinitions) {
                 LogUtil.info("Parsing column definition " + wrapWithQuotes(columnDefinition));
                 String[] parsedColumnDefinition = parseAndReturnColumnDefinition(columnDefinition);
                 table.addColumn(parsedColumnDefinition[0], parsedColumnDefinition[1]);
             }
+//            }
         }
     }
